# backend/app/Http/Requests/Order/RegisterPaymentRequest.php
<?php

namespace App\Http\Requests\Order;

use App\Http\Requests\BaseRequest;
use Illuminate\Validation\Rule;

class RegisterPaymentRequest extends BaseRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
     */
    public function rules(): array
    {
        return [
            'order_id' => ['required', 'exists:orders,id'],
            'payment_type' => ['required', 'string', Rule::in(['cash', 'card'])],
        ];
    }

    public function messages(): array
    {
        return [
            'payment_type.in' => 'Payment type must be either "cash" or "card".',
        ];
    }
}